requirements.txt
email-validator==2.2.0
flask==3.0.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
numpy==2.2.4
opencv-python==4.11.0.86
psycopg2-binary==2.9.10
pytesseract==0.3.13
streamlit==1.44.1
werkzeug==3.0.0
Pillow==10.0.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
python==3.11.7
